moves `Form` into its own component folder and lifts state
